requirements.txt
transformers==4.0.0
sentence-transformers==2.1.0
faiss-cpu==1.7.1
streamlit==1.3.0
pymupdf==1.18.0
torch==1.9.0  # or later for CPU support
openai==0.10.1  # If you're using GPT-3 API
